add cart button to nav

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -48,10 +48,18 @@ const Mutations = {
     );
   },
   async deleteItem(parent, args, ctx, info) {
+      //if they don't have access, catch the error in frontend
     const where = { id: args.id };
     //find the item
-    const item = await ctx.db.query.item({ where }, `{id title}`);
+    const item = await ctx.db.query.item({ where }, `{id title user { id }}`);
     //check if they own that item, or have permish
+    const ownsItem = item.user.id === ctx.request.userId
+    const hasPermissions = ctx.request.user.permissions.some(
+        (permission => ['ADMIN', 'ITEMDELETE'].includes (permission))
+    )
+    if (!ownsItem || !hasPermissions) {
+        throw new Error('You don\'t have permission to do this')
+    }
     //delete
     return ctx.db.mutation.deleteItem({ where }, info); //info: what to return
   },
